\name{ClusterLevel}
\alias{ClusterLevel}

\title{
ClusterLevel
}
\description{
Extract cluster information from given data and cluster information.
}
\usage{
ClusterLevel(data, clusN, clusinf)
}
\arguments{
  \item{data}{Data set to cluster}
  \item{clusN}{The number of clusters}
  \item{clusinf}{The cluster information}
}
\details{
Extract cluster information from given data and cluster information.
}
\value{
Return a vector with the cluster index.
}
\references{
Li, Songzi (2015).
k-groups: A Generalization of k-means by Energy Distance,
Ph.D. thesis, Bowling Green State University.
}
\author{
Songzi Li
}
\keyword{ cluster }
\keyword{ multivariate }%
