\name{Clusters}
\alias{Clusters}

\title{
Clusters
}
\description{
Get cluster information from level.
}
\usage{
Clusters(level)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{level}{A given cluster index vector}
}
\details{
Get cluster information from level.
}
\value{
A list of index vectors for each cluster.
}
\references{
Li, Songzi (2015).
k-groups: A Generalization of k-means by Energy Distance,
Ph.D. thesis, Bowling Green State University.
}
\author{
Songzi Li
}
\keyword{ cluster }
\keyword{ multivariate }
