\name{kgroups-package}
\alias{kgroups-package}
\docType{package}
\title{K-groups Clustering
}
\description{
Implentation of the energy-distance based k-groups clustering algorithms.
}
\details{
\tabular{ll}{
Package: \tab kgroups\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-11-08\cr
License: \tab GPL (>=2.0)\cr
}
}
\author{
Songzi Li and Maria Rizzo
}
\references{
Li, Songzi (2015).
k-groups: A Generalization of k-means by Energy Distance,
Ph.D. thesis, Bowling Green State University.

Songzi Li and Maria L. Rizzo (2017).
K-groups: A Generalization of K-means Clustering, ArXiv e-prints, 1711.04359.
https://arxiv.org/abs/1711.04359
}
\keyword{ package }
\keyword{ cluster }
\keyword{ multivariate }

